feat(api): move copy target profile badge to evaluation context
