+++ api/tests/evaluation/integration/domain/usecases/copy-target-profile-badges_test.js
@@ -1,9 +1,9 @@
-import { copyTargetProfileBadges } from '../../../../lib/domain/usecases/copy-target-profile-badges.js';
-import * as badgeCriteriaRepository from '../../../../src/evaluation/infrastructure/repositories/badge-criteria-repository.js';
-import * as badgeRepository from '../../../../src/evaluation/infrastructure/repositories/badge-repository.js';
-import { withTransaction } from '../../../../src/shared/domain/DomainTransaction.js';
-import { SCOPES } from '../../../../src/shared/domain/models/BadgeDetails.js';
-import { catchErr, databaseBuilder, expect, knex } from '../../../test-helper.js';
+import { copyTargetProfileBadges } from '../../../../../src/evaluation/domain/usecases/copy-target-profile-badges.js';
+import * as badgeCriteriaRepository from '../../../../../src/evaluation/infrastructure/repositories/badge-criteria-repository.js';
+import * as badgeRepository from '../../../../../src/evaluation/infrastructure/repositories/badge-repository.js';
+import { withTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
+import { SCOPES } from '../../../../../src/shared/domain/models/BadgeDetails.js';
+import { catchErr, databaseBuilder, expect, knex } from '../../../../test-helper.js';
 
 describe('Integration | UseCases | copy-badges', function () {
   describe('when there are badges to copy for a target profile', function () {
